navbar created

+++ backend/controllers/product.controller.js
@@ -1,4 +1,5 @@
 import Product from "../models/product.model.js";
+import mongoose from "mongoose";
 
 export const getProducts = async (req, res) => {
   try {
@@ -11,32 +12,54 @@ export const getProducts = async (req, res) => {
 };
 
 export const createProduct = async (req, res) => {
+  const product = req.body;
+  if (!product.name || !product.price || !product.image) {
+    return res
+      .status(400)
+      .json({ success: false, message: "Please enter all the fields" });
+  }
+  const newProduct = new Product(product);
   try {
-    const products = await Product.find({});
-    res.status(200).json({ success: true, data: products });
+    await newProduct.save();
+    res.status(201).json({ success: true, data: newProduct });
   } catch (error) {
-    console.log("Error in fetching products:", error.message);
+    console.log("Error in create product:", error.message);
     res.status(500).json({ success: false, message: "server error" });
   }
 };
 
 export const updateProduct = async (req, res) => {
+  const { id } = req.params;
+  const product = req.body;
+
+  if (!mongoose.Types.ObjectId.isValid(id)) {
+    return res
+      .status(404)
+      .json({ success: false, message: `No product with id: ${id}` });
+  }
+
   try {
-    const products = await Product.find({});
-    res.status(200).json({ success: true, data: products });
+    const updatedProduct = await Product.findByIdAndUpdate(id, product, {
+      new: true,
+    });
+    res.status(200).json({ success: true, data: updatedProduct });
   } catch (error) {
-    console.log("Error in fetching products:", error.message);
+    console.log("Error in updating product:", error.message);
     res.status(500).json({ success: false, message: "server error" });
   }
 };
 
 export const deleteProduct = async (req, res) => {
   const { id } = req.params;
+
+  if (!mongoose.Types.ObjectId.isValid(req.params.id)) {
+    return res.status(404).send(`No product with id: ${req.params.id}`);
+  }
   try {
     await Product.findByIdAndDelete(id);
     res.status(200).json({ success: true, message: "Product deleted" });
   } catch (error) {
     console.error("Error deleting product:", error.message);
-    res.status(404).json({ success: false, message: "Product Not Found" });
+    res.status(500).json({ success: false, message: "Server error" });
   }
 };
